# [libc] Optimize mempcy size thresholds