@dvyukov 2023-10-28
I've updated the change to use the same approach as memmove.
PTAL